Add basic intoto apt transport method (http relay)
Basic in-toto apt transport that forwards messages from apt to the
builtin http transport and vice-versa, providing stubs to intercept
certain messages to perform in-toto verification for the debian
package to be installed. See `README.md` and the extensive code
docs and comments in `intoto.py` for further details.

This is a squashed commit, the WIP-trial-and-error git history is
available in
https://github.com/lukpueh/apt-transport-in-toto/tree/basic-transport-wip.