Add test to mock basic apt<->intoto<->http flow
The test setup provides an apt process (the unittest) that runs the
intoto transport in a subprocess, which in turn calls the http
transport also in a subprocess.

The unit test mocks the apt superprocess. A mock http transport is
added as executable in the same directory as the intoto transport.

Currently the TestCase does not make any explicit assertions, it
just runs through a regular message exchange.