Add test layout, links and final product
Layout is signed with alice' test key
88876A89E3D4698F83D3DB0E72E33CA3E0E04E46

Links are signed with Bobby's and Carly's test keys
5863835E5EC8E640FA24410F069EDC1D59B58507
E946FC6076D683584E6803DBD35BAD6A79A3C6B3

The latter are available (public parts) as functionaries keys in
the layout.

The layout has one step that requires at least two rebuilder link
files and a dummy inspection to link the final product to the
step. The recorded product in the link files is the final product
added in this commit.