Factor out and update handle config message
Create helper function _intoto_parse_config to parse config
items upon reception of 601 Configuration.

This commit also makes the config parser more robust and, in
addition to the existing config items, it looks for a LogLevel
configuration and sets the logging StreamHandler.

This commit also removes the handling of 600 URI Acquire, which
was used previously to detect if a downloaded file is a package
(i.e. not an index file).
Since other files may be acquired that are neither a package nor
an index file, we now base the decision of whether we should do
in-toto verification on the file name format.