Add test files
Add root layout (tests/data/intoto.conf.docker) with one step, i.e.
rebuild and a dummy inspection.
The authorized keys for the rebuild step are the keys that
are used to sign metadata on the currently available public
rebuilders, whose addresses are specified in the added intoto apt
configuration file (tests/data/intoto.conf.docker).
The layout is signed with alice's private
key, which is available in the test gpg keyring
(tests/data/gpg_keyring) under the keyid
88876A89E3D4698F83D3DB0E72E33CA3E0E04E46, and whose public part
is also added in this commit (tests/data/alice.asc).