Add NoFail setting to continue installation on fail
If set to "true" NoFail allows a client to install packages
even if in-toto verification fails, but only if the fail reason
is missing link metadata.

This setting should be removed, once there is broader support for
rebuilder in-toto link metadata.

This commit configures the online demo (see Dockerfile) to enable
the NoFail setting.