Add unit tests and small transport fixes
Huge parts of the in-toto transport are tested in tests_intoto.py
by mocking the entire installation routines. For some special
cases, especially error handling it is not feasible to test them
in the course of an entire installation. Those are covered in
unit tests added with this commit.

This commit also adds to small fixes:
 - Error handling in message deserialization (log level)
 - Remove trailing whitespace in error message.