Rename test layouts
Rename unsecure offline and online (docker) test layouts
to better distinguish them from the included generic layout that
may be used for package verification.

The test layouts are signed with a test key that is included in
this repo. Hence they should not be use used to verify any package.