Add exception handling for BrokenPipeError
If we write to the pipe, used to communicate with the apt parent
process, after apt has closed it, Python raises a BrokenPipeError.

This commit adds exception handling to the write function and a
global boolean BROKENPIPE used as exit condition for the reader
threads.

The commit leaves a few TODO notes that suggest better (more
readable) exception handling.